Kamikaze enemy created!
Kamikaze exploding enemy now created.
Kamikaze enemy will roll at player until reaches explode start radius, initiating explode process.
Kamikaze enemy will damage any enemy within explosion -damage- radius on explosion.

+++ Team_Dragon_Zombies/Assets/Scripts/FriendlyAI.cs
@@ -13,7 +13,7 @@ public class FriendlyAI : MonoBehaviour, IDamage, IFriendly
 
     [Header("-----Stats-----")]
     public int HP;
-    public int faceTargetSpeed;
+    public float faceTargetSpeed;
     public float detectionRadius = 10f;
     public int animSpeedTrans;
     public float attackRate = 1.5f;
